003-stack_cloud_functions-main/cf_csv_to_parquet/requirements.txt
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-logging
fsspec
gcsfs
pyarrow